Fix Buy/Sell buttons UI in Coin page
