+++ walletkit/src/main/java/io/horizontalsystems/walletkit/modules/coin/overview/ui/CoinOverviewScreen.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -25,6 +26,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
+import io.horizontalsystems.marketkit.models.FullCoin
+import io.horizontalsystems.marketkit.models.LinkType
+import io.horizontalsystems.marketkit.models.Token
 import io.horizontalsystems.walletkit.R
 import io.horizontalsystems.walletkit.core.alternativeImageUrl
 import io.horizontalsystems.walletkit.core.iconPlaceholder
@@ -43,13 +47,17 @@ import io.horizontalsystems.walletkit.modules.coin.overview.CoinOverviewViewMode
 import io.horizontalsystems.walletkit.modules.coin.overview.HudMessageType
 import io.horizontalsystems.walletkit.modules.coin.ui.CoinScreenTitle
 import io.horizontalsystems.walletkit.modules.enablecoin.restoresettings.RestoreSettingsViewModel
+import io.horizontalsystems.walletkit.modules.evmfee.ButtonsGroupWithShade
 import io.horizontalsystems.walletkit.modules.managewallets.ManageWalletsModule
 import io.horizontalsystems.walletkit.modules.managewallets.ManageWalletsViewModel
 import io.horizontalsystems.walletkit.modules.markdown.MarkdownPage
+import io.horizontalsystems.walletkit.modules.multiswap.SwapPage
 import io.horizontalsystems.walletkit.modules.nav3.HSNavigation
 import io.horizontalsystems.walletkit.modules.restoreconfig.BirthdayHeightConfigPage
 import io.horizontalsystems.walletkit.ui.compose.ComposeAppTheme
 import io.horizontalsystems.walletkit.ui.compose.HSSwipeRefresh
+import io.horizontalsystems.walletkit.ui.compose.components.ButtonPrimaryDefault
+import io.horizontalsystems.walletkit.ui.compose.components.ButtonPrimaryYellow
 import io.horizontalsystems.walletkit.ui.compose.components.ButtonSecondaryCircle
 import io.horizontalsystems.walletkit.ui.compose.components.ButtonSecondaryDefault
 import io.horizontalsystems.walletkit.ui.compose.components.CellFooter
@@ -62,13 +70,13 @@ import io.horizontalsystems.walletkit.ui.compose.components.body_leah
 import io.horizontalsystems.walletkit.ui.compose.components.subhead2_grey
 import io.horizontalsystems.walletkit.ui.helpers.LinkHelper
 import io.horizontalsystems.walletkit.ui.helpers.TextHelper
-import io.horizontalsystems.marketkit.models.FullCoin
-import io.horizontalsystems.marketkit.models.LinkType
 
 @Composable
 fun CoinOverviewScreen(
     fullCoin: FullCoin,
-    navigation: HSNavigation
+    navigation: HSNavigation,
+    coinToken: Token?,
+    popularToken: Token?
 ) {
     val vmFactory by lazy { CoinOverviewModule.Factory(fullCoin) }
     val viewModel = viewModel<CoinOverviewViewModel>(factory = vmFactory)
@@ -137,137 +145,158 @@ fun CoinOverviewScreen(
 
                     ViewState.Success -> {
                         overview?.let { overview ->
-                            Column(modifier = Modifier.verticalScroll(rememberScrollState())) {
-                                CoinScreenTitle(
-                                    fullCoin.coin.name,
-                                    overview.marketCapRank,
-                                    fullCoin.coin.imageUrl,
-                                    fullCoin.coin.alternativeImageUrl,
-                                    fullCoin.iconPlaceholder
-                                )
+                            Column {
+                                Column(
+                                    modifier = Modifier
+                                        .weight(1f)
+                                        .verticalScroll(rememberScrollState())
+                                ) {
+                                    CoinScreenTitle(
+                                        fullCoin.coin.name,
+                                        overview.marketCapRank,
+                                        fullCoin.coin.imageUrl,
+                                        fullCoin.coin.alternativeImageUrl,
+                                        fullCoin.iconPlaceholder
+                                    )
 
-                                Chart(chartViewModel = chartViewModel)
+                                    Chart(chartViewModel = chartViewModel)
 
-                                Spacer(modifier = Modifier.height(12.dp))
+                                    Spacer(modifier = Modifier.height(12.dp))
 
-                                CellUniversalLawrenceSection {
-                                    RowUniversal(
-                                        modifier = Modifier
-                                            .fillMaxWidth()
-                                            .padding(horizontal = 16.dp),
-                                    ) {
-                                        subhead2_grey(text = stringResource(R.string.CoinPage_Indicators))
-                                        Spacer(modifier = Modifier.weight(1f))
+                                    CellUniversalLawrenceSection {
+                                        RowUniversal(
+                                            modifier = Modifier
+                                                .fillMaxWidth()
+                                                .padding(horizontal = 16.dp),
+                                        ) {
+                                            subhead2_grey(text = stringResource(R.string.CoinPage_Indicators))
+                                            Spacer(modifier = Modifier.weight(1f))
 
-                                        if (chartIndicatorsState.hasActiveSubscription) {
-                                            if (chartIndicatorsState.enabled) {
-                                                ButtonSecondaryDefault(
-                                                    title = stringResource(id = R.string.Button_Hide),
-                                                    onClick = {
-                                                        viewModel.disableChartIndicators()
+                                            if (chartIndicatorsState.hasActiveSubscription) {
+                                                if (chartIndicatorsState.enabled) {
+                                                    ButtonSecondaryDefault(
+                                                        title = stringResource(id = R.string.Button_Hide),
+                                                        onClick = {
+                                                            viewModel.disableChartIndicators()
 
-                                                        stat(
-                                                            page = StatPage.CoinOverview,
-                                                            event = StatEvent.ToggleIndicators(false)
-                                                        )
-                                                    }
-                                                )
-                                            } else {
-                                                ButtonSecondaryDefault(
+                                                            stat(
+                                                                page = StatPage.CoinOverview,
+                                                                event = StatEvent.ToggleIndicators(
+                                                                    false
+                                                                )
+                                                            )
+                                                        }
+                                                    )
+                                                } else {
+                                                    ButtonSecondaryDefault(
+                                                        modifier = Modifier.height(28.dp),
+                                                        title = stringResource(id = R.string.Button_Show),
+                                                        onClick = {
+                                                            viewModel.enableChartIndicators()
+
+                                                            stat(
+                                                                page = StatPage.CoinOverview,
+                                                                event = StatEvent.ToggleIndicators(
+                                                                    true
+                                                                )
+                                                            )
+                                                        }
+                                                    )
+                                                }
+                                                HSpacer(width = 8.dp)
+                                                ButtonSecondaryCircle(
                                                     modifier = Modifier.height(28.dp),
-                                                    title = stringResource(id = R.string.Button_Show),
-                                                    onClick = {
-                                                        viewModel.enableChartIndicators()
+                                                    icon = R.drawable.ic_setting_20
+                                                ) {
+                                                    navigation.slideFromRight(IndicatorsPage)
 
-                                                        stat(
-                                                            page = StatPage.CoinOverview,
-                                                            event = StatEvent.ToggleIndicators(true)
-                                                        )
-                                                    }
-                                                )
-                                            }
-                                            HSpacer(width = 8.dp)
-                                            ButtonSecondaryCircle(
-                                                modifier = Modifier.height(28.dp),
-                                                icon = R.drawable.ic_setting_20
-                                            ) {
-                                                navigation.slideFromRight(IndicatorsPage)
-
-                                                stat(
-                                                    page = StatPage.CoinOverview,
-                                                    event = StatEvent.Open(StatPage.Indicators)
-                                                )
+                                                    stat(
+                                                        page = StatPage.CoinOverview,
+                                                        event = StatEvent.Open(StatPage.Indicators)
+                                                    )
+                                                }
                                             }
                                         }
                                     }
-                                }
 
-                                if (overview.marketData.isNotEmpty()) {
-                                    Spacer(modifier = Modifier.height(12.dp))
-                                    MarketData(overview.marketData)
-                                }
-
-                                if (overview.roi.isNotEmpty()) {
-                                    Spacer(modifier = Modifier.height(24.dp))
-                                    CellSingleLineClear(borderTop = true) {
-                                        body_leah(text = stringResource(R.string.CoinPage_ROI_Title, viewModel.fullCoin.coin.code))
+                                    if (overview.marketData.isNotEmpty()) {
+                                        Spacer(modifier = Modifier.height(12.dp))
+                                        MarketData(overview.marketData)
                                     }
-                                    Spacer(modifier = Modifier.height(12.dp))
-                                    Roi(overview.roi, navigation)
-                                }
 
-                                viewModel.tokenVariants?.let { tokenVariants ->
-                                    Spacer(modifier = Modifier.height(24.dp))
-                                    TokenVariants(
-                                        tokenVariants = tokenVariants,
-                                        onClickAddToWallet = {
-                                            manageWalletsViewModel.enable(it)
-
-                                            stat(
-                                                page = StatPage.CoinOverview,
-                                                event = StatEvent.AddToWallet
+                                    if (overview.roi.isNotEmpty()) {
+                                        Spacer(modifier = Modifier.height(24.dp))
+                                        CellSingleLineClear(borderTop = true) {
+                                            body_leah(
+                                                text = stringResource(
+                                                    R.string.CoinPage_ROI_Title,
+                                                    viewModel.fullCoin.coin.code
+                                                )
                                             )
-                                        },
-                                        onClickRemoveWallet = {
-                                            manageWalletsViewModel.disable(it)
+                                        }
+                                        Spacer(modifier = Modifier.height(12.dp))
+                                        Roi(overview.roi, navigation)
+                                    }
 
-                                            stat(
-                                                page = StatPage.CoinOverview,
-                                                event = StatEvent.RemoveFromWallet
-                                            )
-                                        },
-                                        onClickCopy = {
-                                            TextHelper.copyText(it)
-                                            HudHelper.showSuccessMessage(view, R.string.Hud_Text_Copied)
+                                    viewModel.tokenVariants?.let { tokenVariants ->
+                                        Spacer(modifier = Modifier.height(24.dp))
+                                        TokenVariants(
+                                            tokenVariants = tokenVariants,
+                                            onClickAddToWallet = {
+                                                manageWalletsViewModel.enable(it)
 
-                                            stat(
-                                                page = StatPage.CoinOverview,
-                                                event = StatEvent.Copy(StatEntity.ContractAddress)
-                                            )
-                                        },
-                                        onClickExplorer = {
-                                            LinkHelper.openLinkInAppBrowser(context, it)
+                                                stat(
+                                                    page = StatPage.CoinOverview,
+                                                    event = StatEvent.AddToWallet
+                                                )
+                                            },
+                                            onClickRemoveWallet = {
+                                                manageWalletsViewModel.disable(it)
 
-                                            stat(
-                                                page = StatPage.CoinOverview,
-                                                event = StatEvent.Open(StatPage.ExternalBlockExplorer)
-                                            )
-                                        },
-                                    )
+                                                stat(
+                                                    page = StatPage.CoinOverview,
+                                                    event = StatEvent.RemoveFromWallet
+                                                )
+                                            },
+                                            onClickCopy = {
+                                                TextHelper.copyText(it)
+                                                HudHelper.showSuccessMessage(
+                                                    view,
+                                                    R.string.Hud_Text_Copied
+                                                )
+
+                                                stat(
+                                                    page = StatPage.CoinOverview,
+                                                    event = StatEvent.Copy(StatEntity.ContractAddress)
+                                                )
+                                            },
+                                            onClickExplorer = {
+                                                LinkHelper.openLinkInAppBrowser(context, it)
+
+                                                stat(
+                                                    page = StatPage.CoinOverview,
+                                                    event = StatEvent.Open(StatPage.ExternalBlockExplorer)
+                                                )
+                                            },
+                                        )
+                                    }
+
+                                    if (overview.about.isNotBlank()) {
+                                        Spacer(modifier = Modifier.height(24.dp))
+                                        About(overview.about)
+                                    }
+
+                                    if (overview.links.isNotEmpty()) {
+                                        Spacer(modifier = Modifier.height(24.dp))
+                                        Links(overview.links) { onClick(it, context, navigation) }
+                                    }
+
+                                    Spacer(modifier = Modifier.height(32.dp))
+                                    CellFooter(text = stringResource(id = R.string.Market_PoweredByApi))
                                 }
-
-                                if (overview.about.isNotBlank()) {
-                                    Spacer(modifier = Modifier.height(24.dp))
-                                    About(overview.about)
+                                coinToken?.let {
+                                    CoinBottomButtons(coinToken, popularToken, navigation)
                                 }
-
-                                if (overview.links.isNotEmpty()) {
-                                    Spacer(modifier = Modifier.height(24.dp))
-                                    Links(overview.links) { onClick(it, context, navigation) }
-                                }
-
-                                Spacer(modifier = Modifier.height(32.dp))
-                                CellFooter(text = stringResource(id = R.string.Market_PoweredByApi))
                             }
                         }
 
@@ -369,5 +398,52 @@ fun Loading() {
             color = ComposeAppTheme.colors.grey,
             strokeWidth = 2.dp
         )
+    }
+}
+
+@Composable
+private fun CoinBottomButtons(
+    coinToken: Token,
+    popularToken: Token?,
+    navigation: HSNavigation
+) {
+    ButtonsGroupWithShade {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 24.dp)
+        ) {
+            ButtonPrimaryYellow(
+                modifier = Modifier.weight(1f),
+                title = stringResource(R.string.CoinPage_Buy),
+                onClick = {
+                    navigation.slideFromRight(
+                        SwapPage(
+                            SwapPage.Input(
+                                tokenIn = popularToken,
+                                tokenOut = coinToken
+                            )
+                        )
+                    )
+                }
+            )
+
+            HSpacer(8.dp)
+
+            ButtonPrimaryDefault(
+                modifier = Modifier.weight(1f),
+                title = stringResource(R.string.CoinPage_Sell),
+                onClick = {
+                    navigation.slideFromRight(
+                        SwapPage(
+                            SwapPage.Input(
+                                tokenIn = coinToken,
+                                tokenOut = popularToken
+                            )
+                        )
+                    )
+                }
+            )
+        }
     }
 }
